Add files via upload
The goal of this project was to build a neural network able to classify which letter of the American Sign Language (ASL) alphabet is being signed, given an image of a signing hand. This project is a first step towards building a possible sign language translator, which can take communications in sign language and translate them into written and oral language. Such a translator would greatly lower the barrier for many deaf and mute individuals to be able to better communicate with others in day to day interactions.
                                                                                                  MODULES USED
•	Tensor flow
Tensor flow, a free and open-source software archive, is used to test differentiable computing and data ow across a range of conditions. It is a symbol of calculation archives that are also utilised by neural network processes in engine literacy. In both research and production, Google employs it.
	 	Keras
Keras is an open-source neural network library written in Python. It is designed to simplify the creation of deep learning models and enables fast experimentation by providing a user-friendly interface and prebuilt models. Keras is built on top of other machine learning frameworks like TensorFlow and Theano, and it allows users to easily switch between these frameworks as needed.
	 	Opencv
OpenCV is an open-source computer vision library that provides various tools and functions for real-time computer vision and image processing tasks, including object detection, tracking, segmentation, and more. It is widely used in the eld of computer vision and machine learning for developing real-world applications.
	 	Numpy:
NumPy is a Python library used for scientific computing, particularly for numerical computations with large, multi-dimensional arrays and matrices. It provides various mathematical functions for linear algebra, Fourier transforms, and more, making it a valuable tool for machine learning and data analysis.